setup-mel-builddir: use the core's setup-builddir
This simplifies our scripts by leveraging upstream. The upstream
messages when local.conf or bblayers.conf are missing are rather more
detailed than our own, but are still effective and true for our builds.
We do end up showing conf-notes.txt before we reconfigure the config
files, but that's minor. This avoids duplicating the copies as well as
the TEMPLATECONF logic.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>