setup-mel-builddir: use pyrex if it and docker are available
This will automatically run bitbake and its associated tools and scripts
in a container using pyrex to wrap the use of docker. This avoids issues
with host dependencies and support and ensures a consistent build
environment. Development will still happen on the host, only builds run
in the container with this approach.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>